+++ datahub-actions/setup.py
@@ -18,7 +18,7 @@ from typing import Dict, Set
 import setuptools
 
 package_metadata: dict = {}
-with open("./src/datahub_actions/__init__.py") as fp:
+with open("./src/datahub_actions/_version.py") as fp:
     exec(fp.read(), package_metadata)
