+++ python-build/generate_release_scripts.py
@@ -18,6 +18,7 @@ class Package:
 
 packages = [
     Package(directory="metadata-ingestion", main_module_name="datahub"),
+    Package(directory="datahub-actions", main_module_name="datahub_actions"),
     Package(
         directory="metadata-ingestion-modules/airflow-plugin",
         main_module_name="datahub_airflow_plugin",
